review: isolate client code.
